requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Faker==16.6.1
filelock==3.9.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.26.5
packaging==23.0
platformdirs==3.0.0
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pwinput==1.0.2
PyJWT==2.6.0
PySimpleGUI==4.60.4
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.6
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.19.0
Werkzeug==2.2.3
yarl==1.8.2
